requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-google-genai
google-generativeai
textblob
nltk
ElevenLabs
fastapi
uvicorn
python-multipart
elevenlabs
